trading_platform/requirements.txt
Django
djangorestframework
psycopg2-binary
kafka-python
python-dotenv
requests
pandas

